temp yeet
I ran this before doing any work on it and it was fine up until 1 am, afterwards it was telling me local host not found so I undid all my changes and it still would not load so i redownloaded the original master and it will not run either. Also I could not open the SQL Server dropdown list no matter how many times i restarted everything. Anyways, I couldnt figure out why it wasnt running for me anymore so i kept on making some changes here and there just to keep going but honestly i dont know if anything worked. I tried to remember to comment on everything that I did. I'll try to load this again in an hour or so and see if it works again.